Add bitfield_append and use it when loading reads and the genome.
While here, a significant bug was discovered. When adding kmers to reads,
we were trying to cache previous kmers in order to avoid duplicates. This
code, however, was erroneous and about 20% of the kmers it was dismissing
were in fact not duplicates at all.

Strangely, this bug only resulted in about a 1.5% degredation in
sensitivity, yet it nearly halved our running time.


git-svn-id: svn+ssh://combine-101/home/svn_bio/trunk@106 450597fd-3c47-0410-84d8-f85f2794411d